# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IRENEXA Platform API",
        "description": "Multi-tenant CRM SaaS API — Leads, Contacts, Companies, Deals, Activities, Workflow automation.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Version 1"
        }
    ],
    "paths": {
        "/leads": {
            "get": {
                "tags": [
                    "Leads"
                ],
                "summary": "List leads for the current tenant",
                "operationId": "da2c01e98fdf8007bde77ec53d48a8a7",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of leads"
                    },
                    "401": {
                        "description": "Not authenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Leads"
                ],
                "summary": "Create a new lead",
                "operationId": "8b779aaccdd96c50d3e4da20e92ac5fc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jane Prospect"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jane@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+91 98765 43210",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lead created"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/leads/{lead}/convert": {
            "post": {
                "tags": [
                    "Leads"
                ],
                "summary": "Convert a lead into a Company, Contact, and Deal",
                "operationId": "4b9c85508119aa800b23258f38e74cc0",
                "parameters": [
                    {
                        "name": "lead",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "company_name",
                                    "deal_title",
                                    "deal_amount"
                                ],
                                "properties": {
                                    "company_name": {
                                        "type": "string",
                                        "example": "Acme Inc"
                                    },
                                    "deal_title": {
                                        "type": "string",
                                        "example": "Annual contract"
                                    },
                                    "deal_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 12000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Deal created from the converted lead"
                    },
                    "403": {
                        "description": "Not allowed to create a deal"
                    },
                    "404": {
                        "description": "Lead not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Send the token from POST /login as: Authorization: Bearer {token}",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Leads",
            "description": "Leads"
        }
    ]
}